move conversion funcs into manual_conversions
